refactor(windows): migrate SystemMenuInterop to CsWin32
Deletes Interop/SystemMenuInterop.cs and replaces it with
Branding/SystemMenuPopup.cs, which calls into CsWin32-generated
user32 entry points. BOOL becomes the strongly-typed Foundation.BOOL
struct, the friendly TrackPopupMenu overload drops the nReserved
parameter, and the friendly ClientToScreen takes a
ref System.Drawing.Point so the int/BOOL disagreement the
hand-written file had to annotate is gone. AppIconBadge import
swap is a single line.
