+++ windows/Ghostty/NativeMethods.txt
@@ -2,10 +2,17 @@
 // docs/superpowers/plans/2026-04-12-pr202-plan.md.
 
 // user32
+ClientToScreen
 GetKeyState
+GetSystemMenu
 MapVirtualKey
 MessageBeep
+SendMessage
 SetClassLongPtr
+SetForegroundWindow
+TrackPopupMenu
+TRACK_POPUP_MENU_FLAGS
+WM_SYSCOMMAND
 
 // gdi32
 CreateSolidBrush
